\name{newscrapeR}
\alias{newscrapeR-package}
\alias{newsscraper}
\docType{package}
\title{
Data Mining of Newspapers
}

\description{
Based on the Yahoo! Query Language, this package allows data mining and processing of news articles from several international newspapers and other online sources.
}
\details{
\tabular{ll}{
Package: \tab newscrapeR\cr
Type: \tab Package\cr
Version: \tab 0.7 \cr
Date: \tab 2013-01-01 \cr
License: \tab GPL-2 \cr
}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
Philipp Burckhardt

Maintainer: Philipp Burckhardt <philipp.burckhardt@mansfield.ox.ac.uk>
}
\references{
Information about the Yahoo! Query Language can be found here: \link{http://developer.yahoo.com/yql/}
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
