\name{newscrapeR}
\alias{newscrapeR-package}
\alias{newsscraper}
\docType{package}
\title{
Data Mining of Newspapers
}

\description{
Based on the Yahoo! Query Language, this package allows data mining and processing of news articles from several international newspapers and other online sources.
}

\details{
\tabular{ll}{
Package: \tab newscrapeR\cr
Type: \tab Package\cr
Version: \tab 0.7 \cr
Date: \tab 2013-01-01 \cr
License: \tab GPL-2 \cr
}
}

\author{
Philipp Burckhardt

Maintainer: Philipp Burckhardt <philipp.burckhardt@mansfield.ox.ac.uk>
}
\references{
Information about the Yahoo! Query Language can be found here: \link{http://developer.yahoo.com/yql/}
}

\keyword{Web scraping, 
         data mining}

